requirements.txt
streamlit
langchain
langchain-community
langchain-openai
youtube-transcript-api
faiss-cpu
tiktoken
python-dotenv
